实现基于Vulkan互操作的FSR2(纹理共享不完善)+模块化本机库+添加SRApi

+++ common/src/main/java/io/homo/superresolution/common/minecraft/GpuTextureAdapter.java
@@ -37,7 +37,7 @@ public class GpuTextureAdapter extends GlTexture {
                 texture.getWidth(),
                 texture.getHeight(),
                 1,
-                texture.handle()
+                (int) texture.handle()
         );
         #endif
         this.texture = texture;
